Version 0.2.0-indev
PenguinRegistryManager now takes in an AbstractPenguinRef and is final.
Added Minotaur to build.gradle to automate Modrinth.
Removed ".1" from the end of the version string to be more SemVer
compliant (and it made no sense to have anyway).
Renamed Ref to PenguinCoreRef.
Added documentation to AbstractPenguinRef.
Added method getRefOrCreate to AbstractPenguinRef.
Created MissingRequiredConstructorException, which is thrown from
getRefOrCreate.